gateware.ports: add `SimulationPort` and `SimulationPlatform`.
The `SimulationPort` class is an implementation of the proposed RFC #69
(amaranth-lang/amaranth#1417, amaranth-lang/rfcs#69).

The `SimulationPlatform` class is a temporary workaround for the lack
of built-in knowledge of `SimulationPort` by `io.{Buffer,FFBuffer}`.

In conjunction with the `PortGroup` class, these classes enable testing
I/O gateware. Eventually they should be integrated with the Glasgow
framework (`access.simulation`), but it is not done at the moment.